Extend MI calls, so that event processing can be related to "current time" Extend the system of Block/Wake-handlers, to use the "current" time.
at wake it (the "current" time) is the upper bound of time for which
all antecedent events have been read (from file descriptors).

In the Block handler the time is current time (before we enter the select() call).
Extend MI calls, so that event processing can be related to "current time"